+++ encore_api_cli/commands/configure.py
@@ -1,6 +1,7 @@
-from typing import Optional
+from typing import Optional, Union
 
 import click
+from click_help_colors import HelpColorsGroup
 from tabulate import tabulate
 
 from ..exceptions import ClickException, SettingsValueError
@@ -11,7 +12,7 @@ from ..state import State, pass_state
 from ..utils import get_settings
 
 
-@click.group()
+@click.group(cls=HelpColorsGroup, help_options_color="cyan")
 def cli() -> None:  # noqa: D103
     pass
 
@@ -24,11 +25,23 @@ def configure(ctx: click.Context, state: State) -> None:
     """Configure your AnyMotion Credentials."""
     if ctx.invoked_subcommand is None:
         settings = get_settings(state.profile, use_env=False)
+
         api_url = click.prompt("AnyMotion API URL", default=settings.api_url)
-        client_id = click.prompt("AnyMotion Client ID", default=settings.client_id)
-        client_secret = click.prompt(
-            "AnyMotion Client Secret", default=settings.client_secret
+        client_id = click.prompt(
+            "AnyMotion Client ID",
+            default=make_hidden(settings.client_id),
+            hide_input=True,
         )
+        client_secret = click.prompt(
+            "AnyMotion Client Secret",
+            default=make_hidden(settings.client_secret),
+            hide_input=True,
+        )
+
+        if isinstance(client_id, HiddenCredential):
+            client_id = client_id.value
+        if isinstance(client_secret, HiddenCredential):
+            client_secret = client_secret.value
 
         try:
             settings.write_config(api_url)
@@ -43,16 +56,11 @@ def configure(ctx: click.Context, state: State) -> None:
 def list(state: State) -> None:
     """Show the configuration you use."""
     settings = get_settings(state.profile)
+
     none = click.style("None", fg="yellow")
+    client_id = make_hidden(settings.client_id, none)
+    client_secret = make_hidden(settings.client_secret, none)
 
-    def hidden_credentials(string: Optional[str]) -> str:
-        if string is None or len(string) == 0:
-            return none
-        else:
-            return string[-4:].rjust(20, "*")
-
-    client_id = hidden_credentials(settings.client_id)
-    client_secret = hidden_credentials(settings.client_secret)
     table = tabulate(
         [
             ["profile", state.profile],
@@ -79,3 +87,31 @@ def clear(state: State) -> None:
     settings = get_settings(state.profile)
     settings.write_config(API_URL)
     settings.write_credentials("", "")
+
+
+class HiddenCredential(str):
+    """Make hidden the credentials value.
+
+    Examples:
+        >>> x = HiddenCredential("abcdefghij")
+        >>> str(x)
+        ****************ghij
+        >>> x.value
+        abcdefghij
+    """
+
+    def __init__(self, value: str):
+        self.value = value
+
+    def __str__(self):
+        return self.value[-4:].rjust(20, "*")
+
+
+def make_hidden(
+    value: Optional[str], default: Optional[str] = None
+) -> Optional[Union[str, HiddenCredential]]:
+    """Get hidden value."""
+    if value:
+        return HiddenCredential(value)
+    else:
+        return default